ngn_wep/dashboard/requirements.txt
Flask==3.0.3
mysql-connector-python==9.1.0
python-dotenv==1.0.1
gunicorn==20.1.0
google-cloud-bigquery==3.10.0
google-cloud-storage==2.14.0
tenacity==8.2.2
pytz==2024.1
pandas==2.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.91.0  # ✅ Python 3.9과 호환되는 버전으로 수정
requests==2.31.0
pyarrow==15.0.0
